fix: delete nested files
