--- search-films/src/views/Home.vue
@@ -1,15 +1,0 @@
-<template>
-  <div class="home">
-    <home-page></home-page>
-  </div>
-</template>
-
-<script>
-import { HomePage } from "@/components/pages";
-export default {
-  name: "Home",
-  components: {
-    HomePage,
-  },
-};
-</script>
